added new download card again...

+++ app/src/main/java/com/deniscerri/ytdlnis/database/dao/FormatDao.kt
@@ -8,7 +8,7 @@ import com.deniscerri.ytdlnis.database.models.ResultItem
 @Dao
 interface FormatDao {
     @Query("SELECT * FROM formats WHERE itemId=:itemId")
-    fun getFormatsByItemId(itemId: Int) : List<Format>
+    fun getFormatsByItemId(itemId: Long) : List<Format>
 
     @Insert(onConflict = OnConflictStrategy.REPLACE)
     suspend fun insert(item: Format)
@@ -20,6 +20,6 @@ interface FormatDao {
     suspend fun deleteAll()
 
     @Query("DELETE FROM formats WHERE itemId=:itemId")
-    suspend fun deleteFromatsByItemId(itemId: Int)
+    suspend fun deleteFromatsByItemId(itemId: Long)
 
 }
